Downloads/BOATFLIX/scraper/requirements.txt
requests
beautifulsoup4
supabase-py
